Defensive programming in all API controllers.
Previously, querying an ID that didn't exist would throw an exception.
Now, instead of using `find', `find_with_id' is used, which returns a
nil when the id isn't found, instead of throwing an exception.

Then, before generating the JSON, we check if the variable is nil. If
so, an appropriate error is returned. Otherwise, we return the JSON as
normal.