Remove HABTM between channels and users.
Previously, to allow private channels, we allowed a channel to have specific
users, and users to have specific channels. This is hardly a good way of doing
it, though, since there is significant repeated information.

In any case, we're getting rid of private channels for now, until we think of a
better way of representing them.