Change from `Message' to `Notification'.
Serious changes in the model. `Message' is now a subclass of `Notification'.
Another subclass is `Survey'. This paves the road for implementing surveys.
NOTE: this only fixes the back-end. Front-end creation and display still need to
be adapted to the new API.

TODO:
  * implement Survey
  * fix front-end
  * write tests

Also opt in to errors in transactional callbacks.