Update Option with validations and `dependent: :destroy'.
Now, when an Option id destroyed, its responses are as well.

Validations ensure that each Option belongs to a Survey, and that `option_guts'
is set. Lastly, `option_guts' is unique, given `notification_id', since it
doesn't make sense to have two options in a survey that is the same.