Use RABL for Notification JSON rendering.
Instead of using `to_json', now everything under Notification is RABL'd. So,
when indexing notifications, we map the whole list to a partial. The partial
renders the common fields, then checks if it's a survey. If so, it renders
the survey's Options and Response by deferring to partials.

When showing a notification, it is much the same.

TODO:
  * RABL everything!
  * Clean up RABL code.
  * Reafactor Option and Response out?