Add Responses Controller.
Responses can now be indexed given a Notification / Survey. They can also be
shown given ID. They can also be created given a Notification / Survey.

There are a few simple checks in place during creation. The survey must exist,
and the user must be authorized to answer. Lastly, the option must belong to the
survey.

If there are other responses to the survey by the current_user, they are
destroyed. Currently, this is done very inefficiently. A Reponse should instead
belong to a Survey, and should be indexed like thus. That way, we can quickly
search for duplicates and destroy them if need be.