Fix timestamps in seeds file.
Previously, the DB seeds file would set the start and end times based on
`DateTime.current'. However, in between setting them, the current time would
change, so the `end_time' would be a millisecond or so later than the
`end_time'. This would cause issues with formatting in different apps.